requirements.txt
Django~=5.0.3
channels~=3.0.4
asgiref~=3.8.1
websocket-client~=1.7.0